Resolvido problema
