+++ src/components/Carrinho.js
@@ -1,18 +1,57 @@
 import React from 'react'
-import CarrinhoLogic from '../logic/carrinho'
 
 
 export default class Carrinho extends React.Component {
 
-    carrinhoLogic = new CarrinhoLogic();
+    ValorTotal = 0;
+    ProdutosNoCarrinho = [];
+
+    removerItem = (itemId) => {
+        let produtos = this.ProdutosNoCarrinho;
+        const tamanhoLista = produtos.length;
+
+        for (let i=0; i<tamanhoLista; i++) {
+            if (produtos[i].id === itemId) {
+                this.ValorTotal -= produtos[i].valor;
+                this.produtos.splice(i, 1);
+            }
+        }
+
+        this.ProdutosNoCarrinho = produtos;
+        return this.ValorTotal;
+    }
+
+
+    getItensUnicos = () => {
+        const itens = new Set(this.ProdutosNoCarrinho);
+        const itensArray = Array.from(itens);
+
+        return itensArray;
+    }
+
+    getQuantItem = (itemId) => {
+        let quant = 0;
+        let item;
+
+        for (item of this.ProdutosNoCarrinho) {
+            if (item.id === itemId) {
+                quant++;
+            }
+        }
+
+        return quant;
+    }
 
     render() {
 
+
+
+       
         return(
             <div>
             <h2>CARRINHO</h2>
             {}
-            <p>Total: ₿ {this.carrinhoLogic.ValorTotal.toFixed(2)}</p>
+            {/* <p>Total: ₿ {this.carrinhoLogic.ValorTotal.toFixed(2)}</p> */}
             </div>
         )
     }
